requirements.txt
asgiref==3.7.2
boto3==1.34.63
botocore==1.34.63
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-bootstrap3==23.6
django-paypal==2.0
django-storages==1.14.2
idna==3.6
jmespath==1.0.1
mysqlclient==2.2.4
pillow==10.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
